requirements.txt
crewai
pydantic
python-dotenv
fastapi
uvicorn
